requirements.txt
psycopg2-binary
requests
sqlalchemy
pandas
python-dotenv
